Update obstacle.py
Created a base obstacle class and modified current black-hole class to incorporate inheritance of base class. 
Created a new obstacle called Asteroids, in the base class added a collision function that returns true if element touches obstacle. Also in base class added movement for the obstacles that move the obstacle around the display.